+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 
 
+## [0.11.1](https://github.com/Blobfolio/utc2k/releases/tag/v0.11.1) - TBD
+
+### Changed
+
+* `FmtUtf2k::date` is now const
+* `FmtUtf2k::time` is now const
+* `FmtUtf2k::year` is now const
+
+
+
 ## [0.11.0](https://github.com/Blobfolio/utc2k/releases/tag/v0.11.0) - 2024-10-25
 
 ### New